データタイプで処理を分けるようにファイル分割
